refactor column and trigger into structs
