+++ mydb/table/column.js
@@ -1,3 +1,4 @@
+import Struct from '../helpers/struct.js'
 import { typesArray } from '../types.js'
 
 const isValidName = (name) => name && !name.match(/[`!@#$%^&*()+\-=\[\]{};':"\\|,.<>\/?~]/g)
@@ -13,26 +14,6 @@ const base = {
 
 const allowed = Object.keys(base)
 
-// columns look like { name, type, indexed, unique }
-class Column {
-  #props
-  constructor(props) {
-    const validatedProps = validateProps(props)
-    this.#setProps({ ...base, ...validatedProps })
-  }
-
-  #setProps(props) {
-    allowed.forEach((field) => this[`#${field}`] = props[field])
-  }
-
-  getProps() {
-    if (this.#props) return this.#props
-    const acc = {}
-    allowed.forEach((field) => acc[field] = this[`#${field}`])
-    return this.#props = acc
-  }
-}
-
 const validateProps = (column) => {
   const { name: n, type: t, indexed, unique } = column
   const name = n.toLowerCase()
@@ -42,6 +23,13 @@ const validateProps = (column) => {
   if (!type || !isValidType(type)) throw new Error('Invalid column type: ', name, type)
 
   return { name, type, indexed, unique }
+}
+
+// columns look like { name, type, indexed, unique }
+class Column extends Struct {
+  constructor(props) {
+    super({ props, validateProps, allowed, base })
+  }
 }
 
 export default Column
